// src/_common/database/database-repository.enum.ts
export enum Repositories {
  UsersRepository = 'UsersRepository',
  UserVerificationCodesRepository = 'UserVerificationCodesRepository',
  SecurityGroupsRepository = 'SecurityGroupsRepository',
  AppConfigurationsRepository = 'AppConfigurationsRepository',
  NotificationsRepository = 'NotificationsRepository',
  BoardNotificationsRepository = 'BoardNotificationsRepository',
  NotificationUserStatusesRepository = 'NotificationUserStatusesRepository',
  FilesRepository = 'FilesRepository',
  BankInformationsRepository = 'BankInformationsRepository',
  ContactMessagesRepository = 'ContactMessagesRepository',
  FaqsRepository = 'FaqsRepository',
  UserStoredPaymentMethodRepository = 'UserStoredPaymentMethodsRepository',
  UserSessionsRepository = 'UserSessionsRepository',
  UserSocialAccountsRepository = 'UserSocialAccountsRepository',
  ActiveUsersHistoriesRepository = 'ActiveUsersHistoriesRepository',
  JobTitlesRepository = 'JobTitlesRepository',
  FieldOfTrainingsRepository = 'FieldOfTrainingsRepository',
  LecturerFieldOfTrainingsRepository = 'LecturerFieldOfTrainingsRepository',
  LecturersRepository = 'LecturersRepository',
  LecturerRequestsRepository = 'LecturerRequestsRepository',
  ToolsRepository = 'ToolsRepository',
  CategoriesRepository = 'CategoriesRepository',
  SkillsRepository = 'SkillsRepository',
  CoursesRepository = 'CoursesRepository',
  CourseDetailsRepository = 'CourseDetailsRepository',
  CourseSkillsRepository = 'CourseSkillsRepository',
  CourseToolsRepository = 'CourseToolsRepository',
  SectionsRepository = 'SectionsRepository',
  LessonsRepository = 'LessonsRepository',
  BlogsRepository = 'BlogsRepository',
  BlogCategoriesRepository = 'BlogCategoriesRepository',
  TagsRepository = 'TagsRepository',
  BlogTagsRepository = 'BlogTagsRepository',
  CollectionsRepository = 'CollectionsRepository',
  ChangeLogsRepository = 'ChangeLogsRepository',
  UsersAssignmentsRepository = 'UsersAssignmentsRepository',
  ReviewsRepository = 'ReviewsRepository',
  WorkShopsRepository = 'WorkShopsRepository',
  ContentReportsRepository = 'ContentReportsRepository',
  UserLessonProgressRepository = 'UserLessonProgressesRepository',
  CertificationsRepository = 'CertificationsRepository',
  CommentsRepository = 'CommentsRepository',
  CommentLikesRepository = 'CommentLikesRepository',
  CancelCoursePurchaseRepository = 'CancelCoursePurchasesRepository',
  TransactionsRepository = 'TransactionsRepository',
  LearningProgramRevenueSharesRepository = 'LearningProgramRevenueSharesRepository',
  WalletsRepository = 'WalletsRepository',
  CouponsRepository = 'CouponsRepository',
  CartsRepository = 'CartsRepository',
  CartItemsRepository = 'CartItemsRepository',
  TransactionLogsRepository = 'TransactionLogsRepository',
  PurchasesRepository = 'PurchasesRepository',
  PurchaseItemsRepository = 'PurchaseItemsRepository',
  DiplomasRepository = 'DiplomasRepository',
  DiplomaDetailsRepository = 'DiplomaDetailsRepository',
  DiplomaCoursesRepository = 'DiplomaCoursesRepository',
  DiplomasUsersRepository = 'DiplomasUsersRepository',
  LogsRepository = 'LogsRepository',
  SiteNotificationsRepository = 'SiteNotificationsRepository',
  UserSiteNotificationsRepository = 'UserSiteNotificationsRepository',
  DeletedUsersLogsRepository = 'DeletedUsersLogsRepository',
  WALLET_TRANSACTIONS_REPOSITORY = 'WalletTransactionsRepository',
  DeletedLecturerRequestsLogsRepository = 'DeletedLecturerRequestsLogsRepository',
  SystemConfigRepository = 'SystemConfigsRepository',
  DashboardEmailsRepository = 'DashboardEmailsRepository',
  PolicyRepository = 'PoliciesRepository',
  BlogLikeRepository = 'BlogLikesRepository',
  BlogViewRepository = 'BlogViewsRepository',
  CourseLecturersRepository = 'CourseLecturersRepository',
  // QuizzesRepository = 'QuizzesRepository',
  QuizQuestionsRepository = 'QuizQuestionsRepository',
  QuizAnswersRepository = 'QuizAnswersRepository',
  UserAnswersRepository = 'UserAnswersRepository',
  UserQuizAttemptsRepository = 'UserQuizAttemptsRepository',
  UserAnswerQuizAnswersRepository = 'UserAnswerQuizAnswersRepository',
  UserLessonVisitsRepository = 'UserLessonVisitsRepository',
  UserPasswordHistoriesRepository = 'UserPasswordHistoriesRepository',
  UserCourseDiplomaViewsRepository = 'UserCourseDiplomaViewsRepository',
  SearchKeywordsRepository = 'SearchKeywordsRepository',
}